GameLogic to Render
